+++ classes/forms/class.ilMumieTaskLPSettingsFormGUI.php
@@ -13,6 +13,10 @@
  */
 class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
 {
+    public const DEADLINE_MODE_NONE = 'none';
+    public const DEADLINE_MODE_FIXED = 'fixed';
+    public const DEADLINE_MODE_TIMELIMIT = 'timelimit';
+
     private ilMumieTaskI18N $i18N;
 
     public function __construct($disable_grade_pool_selection)
@@ -25,7 +29,9 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
     private $modus_item;
     private $gradepool_item;
     private $passing_threshold_item;
+    private $deadline_mode_item;
     private $deadline_item;
+    private $timelimit_item;
     private $disable_grade_pool_selection;
 
     public function setFields()
@@ -61,11 +67,28 @@ class ilMumieTaskLPSettingsFormGUI extends ilPropertyFormGUI
         $this->addItem($this->passing_threshold_item);
         $this->passing_threshold_item->setInfo($this->i18N->txt('frm_passing_grade_desc'));
 
+        $this->deadline_mode_item = new ilRadioGroupInputGUI($this->i18N->txt('frm_deadline_mode'), 'deadline_mode');
+        $this->deadline_mode_item->setInfo($this->i18N->txt('frm_deadline_mode_desc'));
+
+        $deadline_mode_option_none = new ilRadioOption($this->i18N->txt('frm_deadline_mode_none'), self::DEADLINE_MODE_NONE);
+        $this->deadline_mode_item->addOption($deadline_mode_option_none);
+
         $this->deadline_item = new ilDateTimeInputGUI($this->i18N->txt('frm_grade_overview_list_deadline'), 'deadline');
         $this->deadline_item->setInfo($this->i18N->txt('frm_lp_deadline_desc'));
         $this->deadline_item->setShowTime(true);
+        $deadline_mode_option_fixed = new ilRadioOption($this->i18N->txt('frm_deadline_mode_fixed'), self::DEADLINE_MODE_FIXED);
+        $deadline_mode_option_fixed->addSubItem($this->deadline_item);
+        $this->deadline_mode_item->addOption($deadline_mode_option_fixed);
 
-        $this->addItem($this->deadline_item);
+        $this->timelimit_item = new ilDurationInputGUI($this->i18N->txt('frm_timelimit'), 'timelimit');
+        $this->timelimit_item->setInfo($this->i18N->txt('frm_timelimit_desc'));
+        $this->timelimit_item->setShowHours(true);
+        $this->timelimit_item->setShowMinutes(true);
+        $deadline_mode_option_timelimit = new ilRadioOption($this->i18N->txt('frm_deadline_mode_timelimit'), self::DEADLINE_MODE_TIMELIMIT);
+        $deadline_mode_option_timelimit->addSubItem($this->timelimit_item);
+        $this->deadline_mode_item->addOption($deadline_mode_option_timelimit);
+
+        $this->addItem($this->deadline_mode_item);
     }
 
     private function getGradepoolInfo()