+++ templates/class.ilMumieTaskTemplateEngine.php
@@ -65,7 +65,7 @@ class ilMumieTaskTemplateEngine
 
     private static function getDeadlineExtensionInformation(ilObjMumieTask $mumie_task, $user_id): string
     {
-        if (ilMumieTaskDeadlineExtensionService::hasDeadlineExtension($user_id, $mumie_task) && $mumie_task->hasDeadline()) {
+        if (ilMumieTaskDeadlineExtensionService::hasDeadlineExtension($user_id, $mumie_task) && $mumie_task->requiresDeadlineSignature()) {
             return ilMumieTaskDeadlineExtensionService::getDeadlineExtensionDate($user_id, $mumie_task);
         }